Query by normalized instead of name
Querying by name causes a clash in the unique field normalized
when someone deletes a project and recreates it with a differently
cased name. This update will have it search by normalized instead.
However to ensure that an update isn't being applied to a project
with the wrong name we assert that the name matches.